Reimplement lexer and parser to be resilient
This follows the matklad tutorial for resilient parsing so that
our lexer and parser are able to tolerate errors and format around
them. Along the way we switch to a more dynamic CST, and use that
to lower to the FST for formatting decisions at the very end.

I have also added an integration test that verifies our lexer
behaves the same as the official Pact lexer, and I've added
reference lexer and parser files from the Pact 5 repository.

We've lost some feature completeness now that we have such a
barebones parser, but as we rebuild parsing for the Pact
language we'll be able to see the formatting develop too.