Add parser integration test
Introduces a new parser integration test, and moves utilities for
interacting with the Pact CLI to a shared test-utils directory.
The parser test works by providing an original input file and
formatted version of the same file to the Pact parser, and then
verifying that the Pact parser produces the same output for each.