Reimplement lexer and parser to be resilient (#12)
* Reimplement lexer and parser to be resilient

This follows the matklad tutorial for resilient parsing so that
our lexer and parser are able to tolerate errors and format around
them. Along the way we switch to a more dynamic CST, and use that
to lower to the FST for formatting decisions at the very end.

I have also added an integration test that verifies our lexer
behaves the same as the official Pact lexer, and I've added
reference lexer and parser files from the Pact 5 repository.

We've lost some feature completeness now that we have such a
barebones parser, but as we rebuild parsing for the Pact
language we'll be able to see the formatting develop too.

* Add parser integration test

Introduces a new parser integration test, and moves utilities for
interacting with the Pact CLI to a shared test-utils directory.
The parser test works by providing an original input file and
formatted version of the same file to the Pact parser, and then
verifying that the Pact parser produces the same output for each.